+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added support for pre-built business card recognition.
 - Added support for pre-built invoices recognition.
 - Added support for providing locale info when recognizing receipts and business cards. Supported locales include support EN-US, EN-AU, EN-CA, EN-GB, EN-IN.
+- Added support for providing the document language in `StartRecognizeContent` when recognizing a form.
 - Added support to train and recognize custom forms with selection marks such as check boxes and radio buttons. This functionality is only available in train with labels scenarios.
 - Added support to `StartRecognizeContent` to recognize selection marks such as check boxes and radio buttons.
 - Added ability to create a composed model from the `FormTrainingClient` by calling method `StartCreateComposedModel`.